refactor중복card제거및Link추가
