+++ client/pages/card.tsx
@@ -1,9 +1,9 @@
-import Card from "components/Maincard/card";
+import Card from "components/Maincard/Card";
 import ApplyModal from "components/Detail/ApplyModal";
 import DetailButtons from "components/Detail/DetailButtons";
 const Cards = () => {
   return (
-    <div className="mt-12 flex flex-row flex-wrap w-full desktop:w-9/12 desktop:min-w-[1000px] tablet:w-full h-fit">
+    <div className="mt-24 flex flex-row flex-wrap w-full desktop:w-9/12 desktop:min-w-[1000px] tablet:w-full h-fit">
       <Card />
       {/* <ApplyModal></ApplyModal> */}
       {/* <DetailButtons></DetailButtons> */}
